added loveseat creative event and Ilian
Ilian is now the storehouse manager. The only thing I know about him is
that his name is slavic, and that he's in charge of distributing goods
from the community's surplus.

I added a line in the intro to make it so your husband answers the
questions "do you miss them?" before telling when the photo was taken.

There was an extra parenthesis somewhere that I deleted.